Encapsulated the complex logic for switching loading components into LoadButton class
